+++ cmd/root.go
@@ -109,15 +109,15 @@ func initConfig() {
 		}
 	}
 
-	// Add the provided name to the labels map for tagging generated resources
-	labels["name"] = name
-
 	// Prepend provided name with 'sonar-' for ease of identifying Sonar deployments.
 	if name != "" {
 		fullName = fmt.Sprintf("%s-%s", nameStub, name)
 	} else {
 		fullName = nameStub
 	}
+
+	// Add the provided name to the labels map for tagging generated resources
+	labels["name"] = fullName
 
 	// If the user has provided a namespace then validate that it looks sane.
 	if rootCmd.PersistentFlags().Lookup("namespace").Changed {