+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.14.0] - 2026-03-05
+
 ### Added
 
 - Add `exec` command.
@@ -161,7 +163,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial release; rough around the edges but usable.
 
-[Unreleased]: https://github.com/a7d-corp/sonar/compare/v0.13.1...HEAD
+[Unreleased]: https://github.com/a7d-corp/sonar/compare/v0.14.0...HEAD
+[0.14.0]: https://github.com/a7d-corp/sonar/compare/v0.13.1...v0.14.0
 [0.13.1]: https://github.com/a7d-corp/sonar/compare/v0.13.1...v0.13.1
 [0.13.1]: https://github.com/a7d-corp/sonar/compare/v0.13.1...v0.13.1
 [0.13.1]: https://github.com/a7d-corp/sonar/compare/v0.13.0...v0.13.1